Laget service for å hente ut og rekjøre tasks.
